pyproject.toml
[project]
name = "pokeagent-speedrun"
version = "0.1.0"
description = "AI agent that plays Pokémon Emerald using vision-language models"
readme = "README.md"
requires-python = ">=3.10,<3.12"

dependencies = [
    "flask==3.0.2",
    "flask-socketio==5.3.6",
    "python-socketio==5.11.1",
    "python-engineio==4.9.0",
    "numpy==1.26.4",
    "opencv-python-headless>=4.9,<5",
    "pillow==10.2.0",
    "pygame==2.6.1",
    "requests>=2.32.2",
    "python-dotenv==1.0.1",
    "eventlet==0.35.2",
    "gevent==23.9.1",
    "gevent-websocket==0.10.1",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "mgba==0.10.2",
    "pyyaml>=5.1",
    "xxhash",
    "more-itertools",
    "pytz>=2020.1",
    "six>=1.14.0",
    "colorama",
    "grpcio>=1.48.2",
    "openai>=1.109.1",
    "google-genai==1.49.0",
    "google-cloud-aiplatform==1.126.0",
    "cffi>=1.6",
    "mcp>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "langfuse==3.9.0",
    "openinference-instrumentation-google-genai>=0.1.0",
    "openinference-instrumentation-vertexai>=0.1.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "chromadb",
    "langchain>=0.3.0",
    "langchain-community",
    "langchain-openai>=1.0.1",
    "langchain-azure-ai",
    "langchain-google-genai",
    "langchain-chroma",
    "azure-ai-vision-imageanalysis>=1.0.0",
    "redis>=7.0.0",
    "redis-entraid>=1.0.0",
    "azure-identity>=1.20.0",
    "scikit-image"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "pokemon_env", "utils", "server"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
]

[tool.ruff.lint.isort]
known-first-party = ["agent", "pokemon_env", "utils", "server"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
